Moving security test folder and LinqExtensions
